+++ Source/PhotinoAPIResponses.cs
@@ -1,0 +1,9 @@
+
+using System;
+using System.Collections.Generic;
+
+namespace PhotinoNET.Responses {
+    public class JsonResponse: Dictionary<String, dynamic> {
+        public JsonResponse(): base() {}
+    }
+}