+++ Tests/APIs/Counter.cs
@@ -1,24 +1,25 @@
 
 using System;
 using System.Collections.Generic;
+using PhotinoNET.Responses;
 
 
 namespace Photino.NET.API.Tests.APIs {
     public class Counter {
         private Int32 count = 0;
 
-        public Dictionary<String, dynamic> CountUp() {
+        public JsonResponse CountUp() {
             this.count += 1;
 
-            return new Dictionary<String, dynamic>(){ { "count", this.count } };
+            return new JsonResponse{ { "count", this.count } };
         }
 
-        public Dictionary<String, dynamic> CountDown() {
+        public JsonResponse CountDown() {
             if (this.count > 0) {
                 this.count -= 1;
             }
 
-            return new Dictionary<String, dynamic>(){ { "count", this.count } };
+            return new JsonResponse{ { "count", this.count } };
         }
     }
 }